[WAL-1395] fix(mdoc2): scope portrait compatibility to issuer2
