+++ waltid-libraries/credentials/waltid-mdoc-credentials/src/commonMain/kotlin/id/walt/mdoc/doc/MDocNameSpaceBuilder.kt
@@ -1,9 +1,7 @@
 package id.walt.mdoc.doc
 
 import id.walt.mdoc.dataelement.MapElement
-import id.walt.mdoc.dataelement.MapKey
 import id.walt.mdoc.dataelement.json.JsonObjectToCborMappingConfig
-import id.walt.mdoc.dataelement.json.mapPortraitCaptureDate
 import id.walt.mdoc.dataelement.json.toDataElement
 import kotlinx.serialization.json.JsonObject
 
@@ -13,22 +11,18 @@ object MDocNameSpaceBuilder {
         nameSpaceId: String,
         jsonData: JsonObject,
         dataMappingConfig: JsonObjectToCborMappingConfig,
-    ): MDocNameSpace {
-        require(jsonData.keys.containsAll(dataMappingConfig.entriesConfigMap.keys)) {
-            "Json keys specified in JSON object config map must all exist in input JSON object"
-        }
-        return MDocNameSpace(
-            nameSpaceId = nameSpaceId,
-            claimsMap = MapElement(jsonData.entries.associate { (key, value) ->
-                MapKey(key) to (mapPortraitCaptureDate(nameSpaceId, key, value)
-                    ?: dataMappingConfig.entriesConfigMap[key]?.executeMapping(value)
-                    ?: value.toDataElement())
-            }),
-        )
-    }
+    ) = MDocNameSpace(
+        nameSpaceId = nameSpaceId,
+        claimsMap = dataMappingConfig.executeMapping(jsonData),
+    )
 
     fun fromJsonObject(
         nameSpaceId: String,
         jsonData: JsonObject,
-    ) = fromJsonObjectMappingConfig(nameSpaceId, jsonData, JsonObjectToCborMappingConfig(emptyMap()))
+    ) = MDocNameSpace(
+        nameSpaceId = nameSpaceId,
+        claimsMap = jsonData.toDataElement() as MapElement,
+    )
 }
+
+
