+++ waltid-services/waltid-issuer-api/src/main/kotlin/id/walt/issuer/issuance/CIProvider.kt
@@ -26,7 +26,6 @@ import id.walt.mdoc.dataelement.ListElement
 import id.walt.mdoc.dataelement.MapElement
 import id.walt.mdoc.dataelement.MapKey
 import id.walt.mdoc.dataelement.toDataElement
-import id.walt.mdoc.dataelement.json.mapPortraitCaptureDate
 import id.walt.mdoc.dataelement.json.toDataElement
 import id.walt.mdoc.doc.MDocBuilder
 import id.walt.mdoc.mso.DeviceKeyInfo
@@ -529,8 +528,7 @@ open class CIProvider(
                     addItemToSign(
                         nameSpace = namespace.key,
                         elementIdentifier = property.key,
-                        elementValue = mapPortraitCaptureDate(namespace.key, property.key, property.value)
-                            ?: property.value.toDataElement(),
+                        elementValue = property.value.toDataElement(),
                     )
                 }
             }
